requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.4
bson==0.5.10
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
Django==5.2
django-extensions==4.0
django-shell-plus==1.1.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huey==2.5.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
Markdown==3.8
markdown2==2.5.3
msgpack==1.1.0
openai==1.72.0
packaging==24.2
platformdirs==4.3.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tavily-python==0.5.4
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
